Renaming teams to msteams
