+++ libraries/microsoft-agents-hosting-msteams/microsoft_agents/hosting/msteams/channel/route_handlers.py
@@ -3,12 +3,12 @@
 
 """Protocol definition for Teams channel update route handlers."""
 
-from typing import Any, Awaitable, Protocol
+from typing import Awaitable, Protocol
 
 from microsoft_teams.api.models.channel_data import ChannelData
 
-from microsoft_agents.hosting.teams.teams_turn_context import TeamsTurnContext
-from microsoft_agents.hosting.teams.type_defs import _StateContra
+from microsoft_agents.hosting.msteams.teams_turn_context import TeamsTurnContext
+from microsoft_agents.hosting.msteams.type_defs import _StateContra
 
 
 class ChannelUpdateHandler(Protocol[_StateContra]):
