+++ libraries/microsoft-agents-hosting-msteams/microsoft_agents/hosting/msteams/config/config.py
@@ -12,12 +12,12 @@ from microsoft_agents.hosting.core import (
     TurnContext,
 )
 
-from microsoft_agents.hosting.teams.teams_turn_context import TeamsTurnContext
-from microsoft_agents.hosting.teams.type_defs import (
+from microsoft_agents.hosting.msteams.teams_turn_context import TeamsTurnContext
+from microsoft_agents.hosting.msteams.type_defs import (
     _RouteDecorator,
     StateT,
 )
-from microsoft_agents.hosting.teams._utils import _send_invoke_response
+from microsoft_agents.hosting.msteams._utils import _send_invoke_response
 
 from .route_handlers import ConfigHandler
