+++ libraries/microsoft-agents-hosting-msteams/microsoft_agents/hosting/msteams/file_consent/route_handlers.py
@@ -7,8 +7,8 @@ from typing import Awaitable, Protocol
 
 from microsoft_teams.api.models import FileConsentCardResponse
 
-from microsoft_agents.hosting.teams.teams_turn_context import TeamsTurnContext
-from microsoft_agents.hosting.teams.type_defs import _StateContra
+from microsoft_agents.hosting.msteams.teams_turn_context import TeamsTurnContext
+from microsoft_agents.hosting.msteams.type_defs import _StateContra
 
 
 class FileConsentHandler(Protocol[_StateContra]):
