+++ libraries/microsoft-agents-hosting-msteams/microsoft_agents/hosting/msteams/task_module/task_module.py
@@ -11,13 +11,13 @@ from microsoft_agents.activity import ActivityTypes
 
 from microsoft_agents.hosting.core import AgentApplication, RouteRank, TurnContext
 
-from microsoft_agents.hosting.teams.teams_turn_context import TeamsTurnContext
-from microsoft_agents.hosting.teams.type_defs import (
+from microsoft_agents.hosting.msteams.teams_turn_context import TeamsTurnContext
+from microsoft_agents.hosting.msteams.type_defs import (
     CommandSelector,
     _RouteDecorator,
     StateT,
 )
-from microsoft_agents.hosting.teams._utils import (
+from microsoft_agents.hosting.msteams._utils import (
     _match_selector,
     _send_invoke_response,
 )
