+++ libraries/microsoft-agents-hosting-msteams/microsoft_agents/hosting/msteams/team/route_handlers.py
@@ -7,8 +7,8 @@ from typing import Awaitable, Protocol
 
 from microsoft_teams.api.models.channel_data import ChannelData
 
-from microsoft_agents.hosting.teams.teams_turn_context import TeamsTurnContext
-from microsoft_agents.hosting.teams.type_defs import _StateContra
+from microsoft_agents.hosting.msteams.teams_turn_context import TeamsTurnContext
+from microsoft_agents.hosting.msteams.type_defs import _StateContra
 
 
 class TeamUpdateHandler(Protocol[_StateContra]):
