+++ tests/hosting_msteams/test_team_lifecycle.py
