+++ tests/hosting_msteams/test_teams_info.py
